+++ app/uploaders/integral/avatar_uploader.rb
@@ -1,8 +1,6 @@
 module Integral
   # Handles uploading user avatars
   class AvatarUploader < ImageUploader
-    include CarrierWave::MiniMagick
-
     # Provide a default URL as a default if there hasn't been a file uploaded
     def default_url
       ActionController::Base.helpers.asset_path("integral/defaults/user_avatar.jpg")