+++ lib/integral/configuration.rb
@@ -12,6 +12,7 @@ module Integral
                   :facebook_app_id,
                   :twitter_handler,
                   :editor_image_size_limit,
+                  :image_compression_quality,
                   :additional_page_templates
 
 
@@ -36,6 +37,7 @@ module Integral
 
       @frontend_parent_controller = Integral::ApplicationController
       @editor_image_size_limit = [800, 800]
+      @image_compression_quality = 90
       @additional_page_templates = []
     end
   end