+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## NEXT
 - Unsaved changes check now occurs for turbolinks visits and has been added on post and list edit pages
 - Toast notifications now fixed positioned (do not scroll out of view)
-
+- Image processing now takes place in the background (not inline)
 
 ## 0.1.4 (March 19, 2017)
 - Users can now clone lists